pyproject.toml
[project]
name = "airtable-contractor-system"
version = "0.1.0"
description = "Airtable Multi-Table Form + JSON Automation System"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "requests>=2.28.0",
    "python-dotenv>=0.19.0",
    "python-decouple>=3.8",
    "openai>=0.27.0",
    "anthropic>=0.5.0",
    "google-generativeai>=0.3.0",
    "PyPDF2>=3.0.0",
    "tenacity>=9.1.2",
]

[tool.uv]
dev-dependencies = [
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "types-requests>=2.28.0",
]

[tool.ruff]
line-length = 88
select = ["E", "W", "F", "I", "C", "B", "UP"]
ignore = []
exclude = [".venv", "__pycache__", ".git", ".mypy_cache", ".pytest_cache"]

[tool.ruff.mccabe]
max-complexity = 10

[tool.ruff.isort]
known-first-party = ["config", "models", "scripts", "utils", "tests"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true

[[tool.mypy.overrides]]
module = [
    "pytest",
    "pytest_cov",
]
ignore_missing_imports = true
